wkcuber/pyproject.toml
[tool.poetry]
name = "wkcuber"
version = "0.0.0"  # filled by dunamai
description = "Python package to create, cube, and work with webKnossos WKW datasets"
authors = ["scalable minds <hello@scalableminds.com>"]
readme = "README.md"
license = "AGPL-3.0"
repository = "https://github.com/scalableminds/webknossos-libs"
homepage = "https://docs.webknossos.org/wkcuber"
include = ["wkcuber/version.py"]  # included explicitly since it's in .gitignore

[tool.poetry.dependencies]
python = ">=3.7.1,<3.10"
attrs = "^21.1.0"
cattrs = "1.7.1"
cluster_tools = { path = "../cluster_tools/", develop = true }
czifile = "^2019.7.2"
GitPython = "^3.0.5"
imagecodecs = "^2020.5.30"
natsort = "^6.2.0"
nibabel = "^2.5.1"
numpy = "^1.19.5"
pillow = ">=6.2.1,<9.0.0"
requests = "^2.22.0"
scikit-image = "^0.18.3"
scikit-learn = "^0.24.0"
scipy = "^1.6.0"
tifffile = "^2020.11.26"
webknossos = { path = "../webknossos/", develop = true }
wkw = "1.1.11"
zarr = "^2.11.0"
universal-pathlib = "0.0.18"

[tool.poetry.dev-dependencies]
black = "22.3.0"
mypy = "0.910"
pylint = "^2.10.2"
pytest = "^6.2.4"
s3fs = "^2022.3.0"
CairoSVG = "^2.5.2"
svgwrite = "^1.4.2"

[tool.poetry.scripts]
wkcuber = "wkcuber.__main__:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
